Update the example image
I updated it by modifying and running test/visual.js. This makes it really easy
to update it in the future: Just run `npm run visual` and grab a screenshot of
the result!

The above also simplifies the code of test/visual.js and fixes a bug where the
thrown error wasn't shown in the console.

This commit is split out of #7.